Fixes for Win10 and many changes to rounding throughout to avoid truncating to 32-bit ints when 64-bits might be available on some architectures. In particular, round() which returns double has been replaced by ROUND32, returning int, and ROUNDBIG returning intptr_t.
git-svn-id: svn://svn.code.sf.net/p/nyquist/code/trunk/nyquist@246 7a394894-9432-4381-9535-169912d1abbc

+++ nyqsrc/pvshell.c
@@ -100,7 +100,7 @@ void pvshell_fetch(snd_susp_type a_susp, snd_list_type snd_list)
     if (flags & PVSHELL_FLAG_TERMINATE) {
         snd_list_terminate(snd_list);
     } else {
-        snd_list->block_len = n;
+        snd_list->block_len = (short) n;
         susp->susp.current += n;
     }
     /* test for logical stop */
